Create TECHMASTEREVEN_2.ino
Experience the magic of the season with this festive Christmas card! Unwrap the joy as it plays your favorite holiday tunes and showcases delightful visuals on its vibrant OLED display. Share the gift of music and merriment with this unique and interactive holiday greeting. The card features three LEDs, one Oled display screen, a push button, a piezo buzzer, this is all controlled by an "Arduino pro micro" microcontroller, it was initially meant to be controlled with the ﻿ESP32-C3 SUPERMINI﻿﻿  but due to some challenges I switched to the "pro micro". The holiday card play Christmas tones with a piezo buzzer and changes songs using the push button connected to the Arduino and, the tone currently playing is show on the Oled screen It also has a sett of LEDs that blink in sync with the tone that is currently playing